pyproject.toml
[project]
name = "medgemma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless>=4.13.0.92",
    "pillow>=10.0.0",
    "protobuf>=5.28.0",
    "sentencepiece>=0.2.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.52.4,<5.0.0",
]
